Add shouldCleanup methods to zOS Batch and File

+++ galasa-managers-parent/galasa-managers-zos-parent/dev.galasa.zosfile.rseapi.manager/src/main/java/dev/galasa/zosfile/rseapi/manager/internal/RseapiZosFileHandlerImpl.java
@@ -110,7 +110,9 @@ public class RseapiZosFileHandlerImpl implements IZosFileHandler {
 	                if (zosDataset.shouldArchive()) {
 	                    zosDataset.archiveContent();
 	                }
-	                zosDataset.delete();
+	                if (zosDataset.shouldCleanup()) {
+	                	zosDataset.delete();
+	                }
 	            }
 			} catch (ZosDatasetException e) {
 				logger.error("Problem in data set cleanup phase", e);
@@ -127,7 +129,9 @@ public class RseapiZosFileHandlerImpl implements IZosFileHandler {
 	            	if (zosVsamDataset.shouldArchive()) {
 	            		zosVsamDataset.archiveContent();
 	            	}
-	                zosVsamDataset.delete();
+	                if (zosVsamDataset.shouldCleanup()) {
+	                	zosVsamDataset.delete();
+	                }
 	            }
 			} catch (ZosVSAMDatasetException e) {
 				logger.error("Problem in VSAM data set cleanup phase", e);
@@ -145,8 +149,10 @@ public class RseapiZosFileHandlerImpl implements IZosFileHandler {
 					if (zosUnixFile.shouldArchive()) {
 						zosUnixFile.archiveContent();
 					}
-		            zosUnixFile.delete();
-		            zosUnixFile.cleanCreatedPath();
+	                if (zosUnixFile.shouldCleanup()) {
+	                	zosUnixFile.delete();
+			            zosUnixFile.cleanCreatedPath();
+	                }
 	            }
 			} catch (ZosUNIXFileException e) {
 				logger.error("Problem in UNIX file cleanup phase", e);
